Move COM to windows lib.
