+++ pyglet/__init__.py
@@ -119,21 +119,18 @@ options = {
     'debug_trace_depth': 1,
     'debug_trace_flush': True,
     'debug_win32': False,
-    'debug_input': False,
     'debug_x11': False,
     'graphics_vbo': True,
     'shadow_window': True,
     'vsync': None,
     'xsync': True,
     'xlib_fullscreen_override_redirect': False,
-    'darwin_cocoa': True,
     'search_local_libs': True,
     'advanced_font_features': False,
     'headless': False,
     'headless_device': 0,
     'win32_disable_shaping': False,
     'dw_legacy_naming': False,
-    'win32_disable_xinput': False,
     'com_mta': False,
     'osx_alt_loop': False
 }
@@ -160,12 +157,11 @@ _option_types = {
     'xsync': bool,
     'xlib_fullscreen_override_redirect': bool,
     'search_local_libs': bool,
-    'win32_gdi_font': bool,
+    'advanced_font_features': bool,
     'headless': bool,
     'headless_device': int,
     'win32_disable_shaping': bool,
     'dw_legacy_naming': bool,
-    'win32_disable_xinput': bool,
     'com_mta': bool,
     'osx_alt_loop': bool,
 }
