+++ pyglet/media/codecs/wmf.py
@@ -36,7 +36,7 @@ import os
 import platform
 import warnings
 
-from pyglet import com, image
+from pyglet import image
 from pyglet.util import debug_print
 from pyglet.libs.win32 import _kernel32 as kernel32
 from pyglet.libs.win32 import _ole32 as ole32
